scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4==4.14.3",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "curl-cffi==0.15.0",
    "google-auth==2.23.4",
    "google-auth-httplib2==0.1.1",
    "google-auth-oauthlib==1.1.0",
    "greenlet==3.2.4",
    "gspread==6.1.0",
    "jsonparse==0.15.1",
    "markdown-it-py==4.0.0",
    "mdurl==0.1.2",
    "pycparser==3.0",
    "pydantic>=2.13.0",
    "pygments==2.20.0",
    "python-dotenv>=1.2.2",
    "redis==7.4.0",
    "rich==14.3.4",
    "rnet==2.4.2",
    "soupsieve==2.8.3",
    "supabase>=2.28.3",
    "tenacity>=9.1.4",
    "typing-extensions==4.15.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
